refactor: move 2 settings and improve mv3
