+++ packages/mask/background/services/settings/index.ts
@@ -1,0 +1,1 @@
+export { __kv_storage_read__, __kv_storage_write__ } from './kv-storage'
