+++ packages/mask/src/extension/background-script/SettingsService.ts
@@ -21,7 +21,8 @@ import {
     currentTokenPricesSettings,
 } from '../../plugins/Wallet/settings'
 import { Flags, MaskMessages } from '../../../shared'
-import { indexedDB_KVStorageBackend, inMemory_KVStorageBackend } from '../../../background/database/kv-storage'
+
+export * from '../../../background/services/settings'
 
 function create<T>(settings: InternalSettings<T>) {
     async function get() {
@@ -84,20 +85,4 @@ export async function setPluginMinimalModeEnabled(id: string, enabled: boolean) 
     currentPluginMinimalModeNOTEnabled['plugin:' + id].value = !enabled
 
     MaskMessages.events.pluginMinimalModeChanged.sendToAll([id, enabled])
-}
-
-export async function __kv_storage_write__(kind: 'indexedDB' | 'memory', key: string, value: unknown) {
-    if (kind === 'memory') {
-        return inMemory_KVStorageBackend.setValue(key, value)
-    } else {
-        return indexedDB_KVStorageBackend.setValue(key, value)
-    }
-}
-
-export async function __kv_storage_read__(kind: 'indexedDB' | 'memory', key: string) {
-    if (kind === 'memory') {
-        return inMemory_KVStorageBackend.getValue(key)
-    } else {
-        return indexedDB_KVStorageBackend.getValue(key)
-    }
 }
